% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_bias_adjustment.R
\name{cluster_bias_adjustment}
\alias{cluster_bias_adjustment}
\title{Cluster bias correction}
\usage{
cluster_bias_adjustment(sigma2js, cluster_size = 22, icc = 0.2)
}
\arguments{
\item{sigma2js}{A vector of sampling variance estimates not including cluster-level variation}

\item{cluster_size}{Average cluster size}

\item{icc}{Intraclass correlation}
}
\value{
Returns a vector of cluster bias adjusted variance estimates
}
\description{
Function to conduct cluster bias correction of variance estimates
obtained from cluster-studies but do not included cluster-level variation
}
\examples{
cbc_var <- cluster_bias_adjustment(
sigma2js = c(0.04, 0.06, 0.08, 0.1),
cluster_size = 15, icc = 0.15
)

cbc_var


}
