% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_J_CHE.R
\name{find_J_CHE}
\alias{find_J_CHE}
\title{Finding number of studies needed to obtain 80\% power}
\usage{
find_J_CHE(
  mu,
  tau2,
  omega2,
  rho,
  var_df = "RVE",
  sigma2_method = "empirical",
  pilot_data_kjsigma2,
  alpha = 0.05,
  target_power = 0.8,
  iterations = 100,
  seed = NULL,
  interval = c(4, 100),
  extendInt = "yes"
)
}
\arguments{
\item{mu}{Smallest effect size of interest/practical concern}

\item{tau2}{Between-study variance}

\item{omega2}{Within-study variance}

\item{rho}{Sample correlation}

\item{var_df}{Variance and degrees of freedom estimation}

\item{sigma2_method}{Defines how sampling variance estimates are obtained}

\item{pilot_data_kjsigma2}{Pilot data containing vectors of kjs and sigma2js}

\item{alpha}{Specifying the nominal Type 1 error rate}

\item{target_power}{Specifying the target power}

\item{iterations}{Number of iterations (default = 100)}

\item{seed}{Set seed to ensure reproducibility of the iterated power approximations}

\item{interval}{a vector containing the end-points of the
interval to be searched for the root.}

\item{extendInt}{character string specifying if the interval c(lower,upper)
should be extended or directly produce an error when f() does not have
differing signs at the endpoints. The default, "yes",
keeps the search interval and hence produces an error. Can be abbreviated.}
}
\value{
Returns a \code{tibble} with information about the sampling methods, type of model,
the preset significant and power levels, the number of studies needed to ontained 80\% power,
and the number of iterations
}
\description{
Finding number of studies needed to obtain 80\% power
}
\examples{
library(POMADE)
library(dplyr)

dat_kjsigma2j <- select(VWB22_pilot, kj, sigma2j = vg_ms_mean)

find_J_CHE(
  mu = 0.1,
  tau2 = 0.1^2,
  omega2 = 0.25^2,
  rho = 0.7,
  pilot_data_kjsigma2 = dat_kjsigma2j,
  seed = 10052510
)

}
