% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDES_plot.R
\name{MDES_plot}
\alias{MDES_plot}
\title{Minimum Detectable Effect Size (MDES) plot function}
\usage{
MDES_plot(
  J,
  tau2,
  omega2,
  rho,
  model = "CHE",
  var_df = "RVE",
  pilot_data_kjsigma2,
  alpha = 0.05,
  target_power = 0.8,
  iterations = 100,
  seed = NULL,
  interval = c(0, 2),
  extendInt = "no",
  min_ES_pr = 0.1,
  expected_studies = NULL,
  ylab = paste0("Minimum detecable ES (alpha = ", alpha, ", Power = ", target_power,
    ")"),
  xlab = "Number of studies (J)",
  Numbers = TRUE,
  number_size = 2.5,
  color = FALSE,
  color_brewer = FALSE,
  brewer_type = "qual",
  palette = 2,
  scales_free = FALSE,
  grid = TRUE,
  breaks = seq(min(J), max(J), 10),
  limits = c(min(J), max(J)),
  legend_position = "bottom",
  mdes_data = FALSE
)
}
\arguments{
\item{J}{A sequence running from the minimum to maximum number of studies
expected to be found}

\item{tau2}{Varying between-study variance estimates}

\item{omega2}{Varying within-study variance estimates}

\item{rho}{Sample correlation coefficients}

\item{model}{Model (default CHE)}

\item{var_df}{Indicate variance and degrees of freedom estimation
("Model", "Satt", "RVE")}

\item{pilot_data_kjsigma2}{Pilot data including a vector of kjs and sampling variance.
If clustering it presented in studies, this should be reflected int the sampling variance
estimates.}

\item{alpha}{Nominal level of statistical significance}

\item{target_power}{Preset power level}

\item{iterations}{Number of times power should be calculated when based on
either simulated or empirical values for kjs and sample sizes or sampling variance estimates}

\item{seed}{Set seed to ensure reproducibility of the plot}

\item{interval}{a vector containing the end-points of the interval to be searched for the root.}

\item{extendInt}{character string specifying if the interval c(lower,upper)
should be extended or directly produce an error when f() does not have
differing signs at the endpoints. The default, "no", keeps the search
interval and hence produces an error. Can be abbreviated.}

\item{min_ES_pr}{Smallest effect size of practical relevance}

\item{expected_studies}{Range in which the expected number of studies
to be found is present.}

\item{ylab}{Change default ylab setting}

\item{xlab}{Change default xlab setting}

\item{Numbers}{Numbering plot in facet grid (Default = \code{TRUE})}

\item{number_size}{Size of numbers on plots}

\item{color}{If \code{ggplot2} colors should be used}

\item{color_brewer}{Customize color brewer}

\item{brewer_type}{Customize color brewer}

\item{palette}{Customize color brewer}

\item{scales_free}{Free scales in \code{facet_grid}}

\item{grid}{Control background grids}

\item{breaks}{Control breaks}

\item{limits}{Control limits}

\item{legend_position}{Control legend position}

\item{mdes_data}{Returns data behind plot}
}
\value{
Returns a facet grid plot with minimum detectable effect sizes
}
\description{
Minimum Detectable Effect Size (MDES) plot function
}
\examples{
library(POMADE)
library(dplyr)

dat_kjsigma2j <- select(VWB22_pilot, kj, sigma2j = vg_ms_mean)

MDES_plot(

  J = seq(50, 100, 10),
  tau2 = c(0, 0.05, 0.1, 0.2)^2,
  omega2 = c(0.05, 0.15, 0.25, 0.35)^2,
  rho = c(.2, .4, .7, .9),
  pilot_data_kjsigma2 = dat_kjsigma2j,
  iterations = 1, # Default = 100
  seed = 13042022,
  expected_studies = c(66, 86)

)
}
