% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_studies_MADE.R
\name{min_studies_MADE}
\alias{min_studies_MADE}
\title{Finding the Number of Studies Needed to Obtain a Certain Amount of Power}
\usage{
min_studies_MADE(
  mu,
  tau,
  omega,
  rho,
  alpha = 0.05,
  target_power = 0.8,
  d = 0,
  model = "CHE",
  var_df = "RVE",
  sigma2_dist = NULL,
  n_ES_dist = NULL,
  iterations = 100,
  seed = NULL,
  warning = TRUE,
  upper = 100,
  show_lower = FALSE
)
}
\arguments{
\item{mu}{Effect size of practical concern}

\item{tau}{Between-study SD}

\item{omega}{Within-study SD}

\item{rho}{Correlation coefficient between effect sizes from the same study}

\item{alpha}{Level of statistical significance}

\item{target_power}{Specifying the target power}

\item{d}{Contrast value}

\item{model}{Indicates dependent effect size model, either "CHE", "CE" or "MLMA". Default = "CHE".}

\item{var_df}{Indicates the technique used to obtained the sampling variance of mu and the degrees of freedom either "Model", "Satt", or "RVE".
Default = "RVE".}

\item{sigma2_dist}{Distribution of sampling variance estimates from each study. Can take either one value, a function, or vector of plausible values}

\item{n_ES_dist}{Distribution of the number of effect sizes per study. Can take either one value, a function, or vector of plausible values}

\item{iterations}{Number of iteration per condition (default = 100)}

\item{seed}{Set seed to ensure reproducibility of the iterated power approximations}

\item{warning}{Return warning when either sigma2_dist or n_ES_dist are based on balanced assumptions}

\item{upper}{Numerical value containing the upper bound of the interval to be searched for the minimum number of studies.}

\item{show_lower}{Logical value indicating whether to report lower bound of the interval searched for the mnimum number of studies. Default is \code{FALSE}.}
}
\value{
Returns a \code{tibble} with information about the expectation of the effect size of practical concern,
the between-study and within-study variance components,
the sample correlation, the contrast effect, the level of statistical significance, the target power value(s),
the number of studies needed, the number of iterations, the model to handle dependent effect sizes,
and the methods used to obtain sampling variance estimates as well as the number effect sizes per study.
}
\description{
Finding the Number of Studies Needed to Obtain a Certain Amount of Power
}
\examples{
studies_needed <- min_studies_MADE(
 mu = 0.2,
 tau = c(0.1, 0.2),
 omega = 0.25,
 rho = 0.7,
 target_power = .8,

 model = "CHE", # default
 var_df = "RVE", # default

 sigma2_dist = \(x) rgamma(x, shape = 5, rate = 10),
 n_ES_dist = \(x) 1 + stats::rpois(x, 5.5 - 1),
 seed = 10052510
)

studies_needed

}
