% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_studies_width_MADE.R
\name{min_studies_width_MADE}
\alias{min_studies_width_MADE}
\title{Finding the Number of Studies Needed to Obtain a Certain Width of the Confidence Intervals}
\usage{
min_studies_width_MADE(
  mu,
  tau,
  omega,
  rho,
  level = 0.95,
  target_width = 0.2,
  model = "CHE",
  var_df = "RVE",
  sigma2_dist = NULL,
  n_ES_dist = NULL,
  iterations = 100,
  seed = NULL,
  warning = TRUE,
  upper = 100
)
}
\arguments{
\item{mu}{Effect size of practical concern.}

\item{tau}{Between-study SD.}

\item{omega}{Within-study SD.}

\item{rho}{Correlation coefficient between effect sizes from the same study.}

\item{level}{Desired coverage level for confidence intervals.}

\item{target_width}{Specify the target width of the estimated confidence interval}

\item{model}{Indicates dependent effect size model, either "CHE", "CE" or "MLMA". Default = "CHE".}

\item{var_df}{Indicates the technique used to obtained the sampling variance of mu and the degrees of freedom either "Model", "Satt", or "RVE".
Default = "RVE".}

\item{sigma2_dist}{Distribution of sampling variance estimates from each study. Can take either one value, a function, or vector of plausible values.}

\item{n_ES_dist}{Distribution of the number of effect sizes per study. Can take either one value, a function, or vector of plausible values.}

\item{iterations}{Number of iteration per condition (default = 100).}

\item{seed}{Set seed to ensure reproducibility of the iterated power approximations.}

\item{warning}{Return warning when either sigma2_dist or n_ES_dist are based on balanced assumptions.}

\item{upper}{Upper value of the uniroot interval}
}
\value{
Returns a \code{tibble} with information about the effect size of practical concern,
the between-study and within-study variance components,  the sample correlation, the target_width,
the expectation of the number of studies needed to obtained the target width of the confidence interval,
the number of iterations, the model to handle dependent effect sizes,
and the methods used to obtain sampling variance estimates as well as the number effect sizes per study.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
studies_needed_width <-
 min_studies_width_MADE(
  mu = 0.1,
  tau = c(0.1, 0.2),
  omega = 0.1,
  rho = c(0.2, 0.7),
  level = c(0.9, 0.95),
  target_width = 0.4,

  model = "CHE",
  var_df = "RVE",

  sigma2_dist = \(x) rgamma(x, shape = 5, rate = 10),
  n_ES_dist = \(x) 1 + stats::rpois(x, 5.5 - 1),

  iterations = 5, # default = 100
  seed = 10052510,
  warning = TRUE,
  upper = 100
)

studies_needed_width




}
