% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MADE.mdes.R
\name{plot_MADE.mdes}
\alias{plot_MADE.mdes}
\title{Plot function for a 'mdes' object}
\usage{
\method{plot_MADE}{mdes}(
  data,
  v_lines = NULL,
  legend_position = "bottom",
  color = TRUE,
  numbers = TRUE,
  number_size = 2.5,
  numbers_ynudge = NULL,
  caption = TRUE,
  x_lab = "Number of studies (J)",
  x_breaks = NULL,
  x_limits = NULL,
  y_breaks = NULL,
  y_limits = NULL,
  y_expand = NULL,
  warning = TRUE,
  traffic_light_assumptions = NULL,
  es_min = NULL,
  expected_studies = NULL,
  ...
)
}
\arguments{
\item{data}{Data/object for which the plot should be made.}

\item{v_lines}{Optional integer or vector to specify vertical line(s) in facet_grid plot(s) (the default is \code{NULL}).}

\item{legend_position}{Optional character string to specify position of legend (default is "bottom").}

\item{color}{Optional logical to specify coloring of plot (default is \code{TRUE}).}

\item{numbers}{Optional logical to specify numbering of plots (default is \code{TRUE}).}

\item{number_size}{Optional integer to specify the size of the plot numbers (default is 2.5).}

\item{numbers_ynudge}{Optional integer to nudge number on the y-axis (default is \code{NULL}).}

\item{caption}{Optional logical to specify inclusion on caption with detailed information regarding
the given analysis (default is \code{TRUE}).}

\item{x_lab}{Title for the x-axis (default is \code{x_lab = "Number of studies (J)"}).}

\item{x_breaks}{Optional sequence to specify breaks on the x-axis (default is \code{NULL}).}

\item{x_limits}{Optional vector to specify the limits of the x-axis (default is \code{NULL}).}

\item{y_breaks}{Optional sequence to specify breaks on the y-axis (default is \code{NULL}).}

\item{y_limits}{Optional vector to specify the limits of the y-axis (default is \code{NULL}).}

\item{y_expand}{Optional vector to expand the limits of the y-axis (default is \code{NULL}).}

\item{warning}{Optional logical to specify if warnings should be returned (default is \code{TRUE}).}

\item{traffic_light_assumptions}{Optional logical to specify coloring of strips of the facet grids to emphasize
assumptions about the likelihood the given analytical scenario (default is \code{NULL}).
See Vembye, Pustejovsky, & Pigott (In preparation) for further details.}

\item{es_min}{Optional integer or vector to either specify a horizontal line or intervals, indicating the minimum
effect size of (practical) relevance wanted to be found (default is \code{NULL}).}

\item{expected_studies}{Optional vector to specify the interval of studies expected to be found in the literature.
When specified this interval will be shaded across facet_grip plots (default is \code{NULL}).}

\item{...}{Additional arguments.}
}
\value{
A \code{ggplot} facet_grip plot showing the minimum detectable effect size across the expected number of
studies faceted by the between-study and within-study SDs, with different colors, lines, and
shapes corresponding to different values of the assumed sample correlation.
}
\description{
Function to make facet_grid plots for an object of class \code{"mdes"}.
}
\details{
\code{plot_MADE.mdes} returns a facet_grip plot for minimum detectable effect size (mdes) analyses conducted with
\code{mdes_MADE}.

In general, it can be rather difficult to guess/approximate the true model parameters
and sample characteristics a priori. Making only a single approximation of the minimum detectable effect size
can easily be misleading even if the true model and data structure slightly diverge
from the yielded data and model assumptions.
To maximize the informativeness of the minimum detectable effect size approximations,
Vembye, Pustejovsky, & Pigott (In preparation) suggest accommodating the uncertainty of the approximations
by reporting or plotting the estimates across a range of possible scenarios, which can be
done by \code{plot_MADE.mdes}.
}
\examples{

mdes_dat <-
  mdes_MADE(
    J = seq(60, 90, 10),
    tau = c(0, 0.25),
    omega = c(0, 0.1),
    rho = c(0.2, 0.7),
    target_power = .8,
    alpha = 0.05,
    sigma2_dist = \(x) rgamma(x, shape = 5, rate = 10),
    n_ES_dist = \(x) 1 + stats::rpois(x, 5.5 - 1),
    seed = 10052510,
    iterations = 5
  )

plot_mdes <- plot_MADE(data = mdes_dat, expected_studies = c(70, 80))
plot_mdes

}
\references{
Vembye, M. H., Pustejovsky, J. E., & Pigott, T. D. (In preparation).
Conducting power analysis for meta-analysis of dependent effect sizes: Common guidelines
and an Introduction to the POMADE R package.
}
\seealso{
\code{\link{plot_MADE}}
}
