% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MADE.min_studies.R
\name{plot_MADE.min_studies}
\alias{plot_MADE.min_studies}
\title{Plot function for a 'min_studies' object}
\usage{
\method{plot_MADE}{min_studies}(
  data,
  v_lines = NULL,
  legend_position = "bottom",
  color = TRUE,
  numbers = TRUE,
  number_size = 2.5,
  numbers_ynudge = NULL,
  caption = TRUE,
  x_lab = NULL,
  x_breaks = NULL,
  x_limits = NULL,
  y_breaks = NULL,
  y_limits = NULL,
  y_expand = NULL,
  warning = TRUE,
  traffic_light_assumptions = NULL,
  v_shade = NULL,
  h_lines = NULL,
  ...
)
}
\arguments{
\item{data}{Data/object for which the plot should be made.}

\item{v_lines}{Optional integer or vector to specify vertical line(s) in facet_grid plot(s) (the default is \code{NULL}).}

\item{legend_position}{Optional character string to specify position of legend (default is "bottom").}

\item{color}{Optional logical to specify coloring of plot (default is \code{TRUE}).}

\item{numbers}{Optional logical to specify numbering of plots (default is \code{TRUE}).}

\item{number_size}{Optional integer to specify the size of the plot numbers (default is 2.5).}

\item{numbers_ynudge}{Optional integer to nudge number on the y-axis (default is \code{NULL}).}

\item{caption}{Optional logical to specify inclusion on caption with detailed information regarding
the given analysis (default is \code{TRUE}).}

\item{x_lab}{Title for the x-axis (default is \code{NULL}). If \code{length(data$mu) == 1}, the x-axis label will
be "Effect Size Level SD". If \code{length(data$mu) > 1}, the x-axis label will "Effect Size of Practical Concern".}

\item{x_breaks}{Optional sequence to specify breaks on the x-axis (default is \code{NULL}).}

\item{x_limits}{Optional vector to specify the limits of the x-axis (default is \code{NULL}).}

\item{y_breaks}{Optional sequence to specify breaks on the y-axis (default is \code{NULL}).}

\item{y_limits}{Optional vector to specify the limits of the y-axis (default is \code{NULL}).}

\item{y_expand}{Optional vector to expand the limits of the y-axis (default is \code{NULL}).}

\item{warning}{Optional logical to specify if warnings should be returned when multiple models appear in the data
(default is \code{TRUE}).}

\item{traffic_light_assumptions}{Optional logical to specify coloring of strips of the facet grids to emphasize
assumptions about the likelihood the given analytical scenario (default is \code{NULL}).
See Vembye, Pustejovsky, & Pigott (In preparation) for further details.}

\item{v_shade}{Optional vector to specify range of the x-axis interval to be shaded on facet_grip plots.}

\item{h_lines}{Optional integer or vector to specify horizontal lines on plots.}

\item{...}{Additional arguments.}
}
\value{
A \code{ggplot} facet_grip plot showing the minimum studies needed to obtained a given
effect size with a certain amount of power and level-alpha across estimates of the within-study SD
faceted by the between-study SDs, with different colors, lines, and
shapes corresponding to different values of the assumed sample correlation. If \code{length(data$mu) > 1},
it returns a \code{ggplot} facet_grip plot showing the minimum studies needed to obtained a given
effect size with a certain amount of power and level-alpha across effect sizes of practical concern
faceted by the between-study and within-study SDs, with different colors, lines, and
shapes corresponding to different values of the assumed sample correlation.
}
\description{
Function to make facet_grid plots for an object of class \code{"min_studies"}.
}
\details{
\code{plot_MADE.min_studies} returns a facet_grip plot with analyses of the minimum studies needed to obtained a given
effect size with a certain amount of power (usually 80\%) and a pre-specified level-alpha conducted with
\code{min_studies_MADE}.

In general, it can be rather difficult to guess/approximate the true model parameters
and sample characteristics a priori. Making only a single approximation of the minimum studies needed to obtained a given
effect size with a certain amount of power (usually 80\%) and a pre-specified level-alpha can easily be misleading
even if the true model and data structure only slightly diverge
from the yielded data and model assumptions.
To maximize the informativeness of the minimum studies needed to obtained a given
effect size approximations, Vembye, Pustejovsky, & Pigott (In preparation) suggest accommodating the uncertainty
of the approximations by reporting or plotting estimates across a range of possible scenarios,
which can be done by \code{plot_MADE.min_studies}.
}
\examples{

min_studies_obj <-
  min_studies_MADE(
    mu = 0.2,
    tau = c(0.1, 0.2),
    omega = c(0, 0.1, 0.2, 0.3),
    rho = c(0.2, 0.7),
    target_power = .8,
    sigma2_dist = \(x) rgamma(x, shape = 5, rate = 10),
    n_ES_dist = \(x) 1 + stats::rpois(x, 5.5 - 1),
    seed = 10052510,
    iterations = 5

  )

min_studies_example <- plot_MADE(data = min_studies_obj)
min_studies_example

}
\references{
Vembye, M. H., Pustejovsky, J. E., & Pigott, T. D. (In preparation).
Conducting power analysis for meta-analysis of dependent effect sizes: Common guidelines
and an Introduction to the POMADE R package.
}
\seealso{
\code{\link{plot_MADE}}
}
