% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_MADE.R
\name{plot_MADE}
\alias{plot_MADE}
\title{Generic plot function for 'MADE' objects}
\usage{
plot_MADE(
  data,
  v_lines,
  legend_position,
  color,
  numbers,
  number_size,
  numbers_ynudge,
  caption,
  x_lab,
  x_breaks,
  x_limits,
  y_breaks,
  y_limits,
  y_expand = NULL,
  warning,
  traffic_light_assumptions,
  ...
)
}
\arguments{
\item{data}{Data/object for which the plot should be made.}

\item{v_lines}{Optional integer or vector to specify vertical line(s) in facet_grid plot(s).}

\item{legend_position}{Optional character string to specify position of legend.}

\item{color}{Optional logical to specify coloring of plot.}

\item{numbers}{Optional logical to specify numbering of plots.}

\item{number_size}{Optional integer to specify the size of the plot numbers.}

\item{numbers_ynudge}{Optional integer to nudge number on the y-axis.}

\item{caption}{Optional logical to specify inclusion on caption with detailed information regarding
the given analysis.}

\item{x_lab}{Title for the x-axis. If \code{NULL}, the x_lab is specified automatically
within the given S3 method.}

\item{x_breaks}{Optional sequence to specify breaks on the x-axis.}

\item{x_limits}{Optional vector to specify the limits of the x-axis.}

\item{y_breaks}{Optional sequence to specify breaks on the y-axis.}

\item{y_limits}{Optional vector to specify the limits of the y-axis.}

\item{y_expand}{Optional vector to expand the limits of the y-axis.}

\item{warning}{Optional logical to specify if warnings should be returned when multiple models appear in the data.}

\item{traffic_light_assumptions}{Optional logical to specify coloring of strips of the facet grids to emphasize
assumptions about the likelihood the given analytical scenario.
See Vembye, Pustejovsky, & Pigott (In preparation) for further details.}

\item{...}{Additional arguments available for some classes of objects.}
}
\value{
A \code{ggplot} object
}
\description{
This is a generic function to make facet_grid plots, with specific methods defined for
\code{\link{power_MADE}}, \code{\link{mdes_MADE}}, and
\code{\link{min_studies_MADE}} objects.
}
\details{
\code{plot_MADE} returns a facet_grip plot for a set of power analyses.

In general, it can be rather difficult to guess/approximate the true model parameters
and sample characteristics a priori. Making only a single power approximation can easily be misleading
even if the true model and data structure only slightly diverge
from the yielded data and model assumptions.
To maximize the informativeness of the power approximations,
Vembye, Pustejovsky, & Pigott (In preparation) suggest accommodating the uncertainty of the power approximations
by reporting or plotting power estimates across a range of possible scenarios, which can be
done by \code{plot_MADE}.
}
\examples{

power_dat <-
  power_MADE(
    J = seq(40, 60, 5),
    mu = 0.1,
    tau = c(0.05, 0.1, 0.2),
    omega = c(0.1, 0.2),
    rho = c(0.2, 0.7),
    sigma2_dist = \(x) rgamma(x, shape = 5, rate = 10),
    n_ES_dist = \(x) 1 + stats::rpois(x, 5.5 - 1),
    iterations = 5, # default is 100
    seed = 10052510
  )

power_example <-
  plot_MADE(
   data = power_dat,
   power_min = 0.8,
   expected_studies = c(45, 55),
   warning = FALSE,
   caption = TRUE,
   color = TRUE,
   model_comparison = FALSE,
   #traffic_light_assumptions = c("unlikely", "likely", "expected", "likely", "expected")
   )

power_example


}
\references{
Vembye, M. H., Pustejovsky, J. E., & Pigott, T. D. (In preparation).
Conducting power analysis for meta-analysis of dependent effect sizes: Common guidelines
and an Introduction to the POMADE R package.
}
\seealso{
\code{\link{plot_MADE.power}}, \code{\link{plot_MADE.mdes}}, \code{\link{plot_MADE.mdes}}
}
