% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/POMADE-package.R
\docType{package}
\name{POMADE-package}
\alias{POMADE}
\alias{POMADE-package}
\title{POMADE: POwer for Meta-Analysis of Dependent Effects}
\description{
Provides functions to compute and plot power for meta-analysis of dependent effect sizes.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/MikkelVembye/POMADE/issues}
}

}
\author{
\strong{Maintainer}: Mikkel H. Vembye \email{mikkel.vembye@gmail.com} (\href{https://orcid.org/0000-0001-9071-0724}{ORCID})

Authors:
\itemize{
  \item James E. Pustejovsky \email{jepusto@gmail.com} (\href{https://orcid.org/0000-0003-0591-9465}{ORCID})
}

}
\keyword{internal}
