% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_CE.R
\name{power_CE}
\alias{power_CE}
\title{Approximate power for the correlated effects (CE) model}
\usage{
power_CE(
  J,
  tau2,
  omega2,
  beta,
  rho,
  var_df = "RVE",
  sample_size_method = c("balanced", "stylized", "empirical"),
  sigma2_method = c("balanced", "empirical"),
  k_mean = NULL,
  N_mean = NULL,
  N_dist = NULL,
  pilot_data_kjN = NULL,
  sigma2_mean = NULL,
  pilot_data_kjsigma2 = NULL,
  alpha = c(0.01, 0.05, 0.1),
  iterations = 100,
  average_power = TRUE,
  seed = NULL
)
}
\arguments{
\item{J}{Expected number of studies to be found}

\item{tau2}{Between-study variance}

\item{omega2}{Within-study variance}

\item{beta}{Smallest effect size of practical/substantial concern}

\item{rho}{Sample correlation}

\item{var_df}{Variance and degrees of freedom estimation}

\item{sample_size_method}{Defines how sample sizes are obtained}

\item{sigma2_method}{Defines how sampling variance estimates are obtained}

\item{k_mean}{Average number of effect sizes per study}

\item{N_mean}{Average (effective) sample size of study}

\item{N_dist}{Simulated distribution o}

\item{pilot_data_kjN}{Pilot data containing kjs and effective sample sizes}

\item{sigma2_mean}{Average sampling variance}

\item{pilot_data_kjsigma2}{Pilot data containing vectors of kjs and sigma2js}

\item{alpha}{statistical significance level}

\item{iterations}{Number of iterations (default = 100)}

\item{average_power}{Average power across the x iterated power estimations}

\item{seed}{Set seed to ensure reproducibility of the iterated power approximations}
}
\value{
Returns a \code{tibble} with information about the sampling methods, type of model, smallest effect size
considered to be of practical concern, the estimate variance for the effect size, degrees of freedom
the power estimate, and the number of iterations
}
\description{
Approximate power for the correlated effects (CE) model
}
\examples{
balanced_CE_RVE_power <-
 power_CE(
  J = 50,
  tau2 = 0.2^2,
  omega2 = 0.1^2,
  beta = 0.1,
  rho = 0.7,
  var_df = "RVE",
  sample_size_method = "balanced",
  k_mean = 4,
  N_mean = 50,
  alpha = 0.05,
  seed = 240222
 )

balanced_CE_RVE_power

}
