% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_plot.R
\name{power_plot}
\alias{power_plot}
\title{Power plot function}
\usage{
power_plot(
  J,
  tau2,
  omega2,
  beta,
  rho,
  k_mean = NULL,
  model = "CHE",
  var_df = "RVE",
  sigma2_method = "empirical",
  pilot_data_kjsigma2 = NULL,
  alpha = 0.05,
  iterations = 100,
  power_min = 0.8,
  expected_studies = NULL,
  seed = NULL,
  ylab = NULL,
  xlab = "Number of studies (J)",
  Numbers = TRUE,
  number_size = 2.5,
  color = FALSE,
  color_brewer = FALSE,
  brewer_type = "qual",
  palette = 2,
  scales_free = FALSE,
  grid = TRUE,
  breaks = seq(min(J), max(J), 10),
  limits = c(min(J), max(J)),
  legend_position = "bottom"
)
}
\arguments{
\item{J}{Insert}

\item{tau2}{Insert}

\item{omega2}{Insert}

\item{beta}{Insert}

\item{rho}{Insert}

\item{k_mean}{Insert}

\item{model}{Insert}

\item{var_df}{Insert}

\item{sigma2_method}{Insert}

\item{pilot_data_kjsigma2}{Insert}

\item{alpha}{Insert}

\item{iterations}{Insert}

\item{power_min}{Insert}

\item{expected_studies}{Insert}

\item{seed}{Insert}

\item{ylab}{Insert}

\item{xlab}{Insert}

\item{Numbers}{Insert}

\item{number_size}{Insert}

\item{color}{Insert}

\item{color_brewer}{Insert}

\item{brewer_type}{Insert}

\item{palette}{Insert}

\item{scales_free}{Insert}

\item{grid}{Insert}

\item{breaks}{Insert}

\item{limits}{Insert}

\item{legend_position}{Insert}
}
\value{
Returns a facet_grid power plot
}
\description{
Power plot function
}
