% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_MADE.R
\name{precision_MADE}
\alias{precision_MADE}
\title{Precision Analysis for Meta-Analysis of Dependent Effect Sizes}
\usage{
precision_MADE(
  J,
  mu,
  tau,
  omega,
  rho,
  level = 0.95,
  model = "CHE",
  var_df = "RVE",
  sigma2_dist = NULL,
  n_ES_dist = NULL,
  iterations = 100,
  seed = NULL,
  warning = TRUE,
  average_precision = TRUE
)
}
\arguments{
\item{J}{Number of studies.}

\item{mu}{Effect size of practical concern.}

\item{tau}{Between-study SD.}

\item{omega}{Within-study SD.}

\item{rho}{Correlation coefficient between effect sizes from the same study.}

\item{level}{Desired coverage level for confidence intervals.}

\item{model}{Indicates dependent effect size model, either "CHE", "CE" or "MLMA". Default = "CHE".}

\item{var_df}{Indicates the technique used to obtained the sampling variance of mu and the degrees of freedom either "Model", "Satt", or "RVE".
Default = "RVE".}

\item{sigma2_dist}{Distribution of sampling variance estimates from each study. Can take either one value, a function, or vector of plausible values.}

\item{n_ES_dist}{Distribution of the number of effect sizes per study. Can take either one value, a function, or vector of plausible values.}

\item{iterations}{Number of iteration per condition (default = 100).}

\item{seed}{Set seed to ensure reproducibility of the iterated power approximations.}

\item{warning}{Return warning when either sigma2_dist or n_ES_dist are based on balanced assumptions.}

\item{average_precision}{Average \code{"lower_bound"}, \code{"upper_bound"}, and \code{"width"} across the number of iteration for each condition.}
}
\value{
Returns a \code{tibble} with information about the expectation of the number of
studies, the effect size of practical concern, the between-study and within-study variance components,
the sample correlation, Desired coverage level for confidence intervals,
the sampling variance of overall average effect size of practical concern, the degrees of freedom,
the lower bound of the confidence interval, the upper bound of the confidence interval, and the width of confidence interval,
the MCSE of the width, the number of iterations, the model to handle dependent effect sizes,
and the methods used to obtain sampling variance estimates as well as the number effect sizes per study.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
precision <-
 precision_MADE(
   J = c(40, 60),
   mu = 0.1,
   tau = 0.2,
   omega = 0.1,
   rho = 0.7,
   sigma2_dist = \(x) rgamma(x, shape = 5, rate = 10),
   n_ES_dist = \(x) 1 + stats::rpois(x, 5.5 - 1),
   model = c("CHE", "MLMA", "CE"),
   var_df = c("Model", "Satt", "RVE"),
   level = .95,
   seed = 10052510
 )

precision




}
