% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDES_MADE.R
\name{MDES_MADE}
\alias{MDES_MADE}
\title{Minimum Detectable Effect Size (MDES) for Meta-Analysis With Dependent Effect Sizes}
\usage{
MDES_MADE(
  J,
  tau2,
  omega2,
  rho,
  alpha = 0.05,
  target_power = 0.8,
  d = 0,
  model = "CHE",
  var_df = "RVE",
  sigma2_dist = NULL,
  n_ES_dist = NULL,
  iterations = 100,
  seed = NULL,
  warning = TRUE,
  interval = c(0, 2)
)
}
\arguments{
\item{J}{Number of studies}

\item{tau2}{Between-study variance}

\item{omega2}{Within-study variance}

\item{rho}{Correlation coefficient between effect sizes from the same}

\item{alpha}{Level of statistical significance}

\item{target_power}{Specifying the target power}

\item{d}{Contrast value}

\item{model}{Indicates dependent effect size model}

\item{var_df}{Indicates the technique used to obtained the sampling variance of mu and the degrees of freedom}

\item{sigma2_dist}{Distribution of sampling variance estimates from each study. Can take either one value, a function, or vector of plausible values}

\item{n_ES_dist}{Distribution of the number of effect sizes per study. Can take either one value, a function, or vector of plausible values}

\item{iterations}{Number of iteration per condition (default = 100)}

\item{seed}{Set seed to ensure reproducibility of the iterated power approximations}

\item{warning}{Return warning when either sigma2_dist or n_ES_dist are based on balanced assumptions}

\item{interval}{A vector containing the end-points of the interval to be searched for the root}
}
\value{
Returns a \code{tibble} with information about the expectation of the number of
studies, the between-study and within-study variance components,
the sample correlation, the contrast effect, the level of statistical significance, the target power value(s),
the minimum detectable effect size, the number of iterations, the model to handle dependent effect sizes,
and the methods used to obtain sampling variance estimates as well as the number effect sizes per study.
}
\description{
Minimum Detectable Effect Size (MDES) for Meta-Analysis With Dependent Effect Sizes
}
\examples{

mdes <- MDES_MADE(
  J = c(40, 60),
  tau2 = 0.2^2,
  omega2 = 0.1^2,
  rho = 0.7,
  model = c("CHE", "MLMA", "CE"),
  var_df = c("Model", "Satt", "RVE"),
  sigma2_dist = \(x) rgamma(x, shape = 5, rate = 10),
  n_ES_dist = \(x) 1 + stats::rpois(x, 5.5 - 1),
  seed = 10052510
)

mdes


}
