% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_plot.R
\name{power_plot}
\alias{power_plot}
\title{Power plot function}
\usage{
power_plot(
  J,
  tau2,
  omega2,
  beta,
  rho,
  k_mean = NULL,
  model = c("CHE", "MLMA"),
  var_df = "RVE",
  N_mean = NULL,
  pilot_data_kjN = NULL,
  sigma2_mean = NULL,
  pilot_data_kjsigma2 = NULL,
  alpha = 0.05,
  iterations = 100,
  power_min = 0.8,
  expected_studies = NULL,
  seed = NULL,
  ylab = NULL,
  xlab = "Number of studies (J)",
  Numbers = TRUE,
  number_size = 2.5,
  color = FALSE,
  color_brewer = FALSE,
  brewer_type = "qual",
  palette = 2,
  scales_free = FALSE,
  grid = TRUE,
  breaks = seq(min(J), max(J), 10),
  limits = c(min(J), max(J)),
  legend_position = "bottom",
  power_data = FALSE
)
}
\arguments{
\item{J}{A sequence running from the minimum to maximum number of studies
expected to be found}

\item{tau2}{Varying between-study variance estimates}

\item{omega2}{Varying within-study variance estimates}

\item{beta}{Smallest effect size of practical concern. Insert one value only}

\item{rho}{Sample correlation coefficients}

\item{k_mean}{Average number of effect sizes per study}

\item{model}{Model (either CHE, MLMA, or CE)}

\item{var_df}{Indicate variance and degrees of freedom estimation
("Model", "Satt", "RVE")}

\item{N_mean}{Average sample size}

\item{pilot_data_kjN}{Pilot data including a vector of kjs and sample sizes.
If clustering it presented in studies, use effective sample sizes.}

\item{sigma2_mean}{Average sampling variance}

\item{pilot_data_kjsigma2}{Pilot data including a vector of kjs and sampling variance.
If clustering it presented in studies, this should be reflected int the sampling variance
estimates.}

\item{alpha}{Nominal level of statistical significance}

\item{iterations}{Number of times power should be calculated when based on
either simulated or empirical values for kj and sample sizes or sampling variance estimates}

\item{power_min}{Minimum acceptable power level}

\item{expected_studies}{Range in which the expected number of studies
to be found is present.}

\item{seed}{Set seed to ensure reproducibility of the plot}

\item{ylab}{Change default ylab setting}

\item{xlab}{Change default xlab setting}

\item{Numbers}{Numbering plot in facet grid (Default = \code{TRUE})}

\item{number_size}{Size of numbers on plots}

\item{color}{If \code{ggplot2} colors should be used}

\item{color_brewer}{Customize color brewer}

\item{brewer_type}{Customize color brewer}

\item{palette}{Customize color brewer}

\item{scales_free}{Free scales in \code{facet_grid}}

\item{grid}{Control background grids}

\item{breaks}{Control breaks}

\item{limits}{Control limits}

\item{legend_position}{Control legend position}

\item{power_data}{Print approximated data. If \code{TRUE}, it returns the dataset only}
}
\value{
Returns a facet_grid power plot
}
\description{
Power plot function
}
\examples{

library(dplyr)

coteach_dat <- VWB22_pilot
dat_kjsigma2j <- dplyr::select(coteach_dat, kj, sigma2j = vg_ms_mean)

# Color format

power_CHE_RVE_color_plot <-
  power_plot(
    J = seq(50, 100, 10),
    tau2 = c(0, 0.05, 0.1, 0.2)^2,
    omega2 = c(0.05, 0.15, 0.25, .35)^2,
    beta = 0.1,
    rho = c(.2, .4, .7, .9),
    model = "CHE",
    var_df = "RVE",
    pilot_data_kjsigma2 = dat_kjsigma2j,
    expected_studies = c(66, 86),
    color = TRUE,
    color_brewer = TRUE,
    iterations = 1,
    seed = 10052510
  )

power_CHE_RVE_color_plot


}
