% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_light_power_plot.R
\name{traffic_light_power_plot}
\alias{traffic_light_power_plot}
\title{Make traffic light power plot}
\usage{
traffic_light_power_plot(power_plot, assumptions)
}
\arguments{
\item{power_plot}{Facet grid \code{power_plot} object}

\item{assumptions}{Assumption regarding the likelihood of the given scenario
in the \code{power_plot}. Make assumptions from upper-left strip to lower-right strip}
}
\value{
Return a traffic light power plot
}
\description{
Make traffic light power plot
}
