% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traffic_light_power_plot.R
\name{traffic_light_power_plot}
\alias{traffic_light_power_plot}
\title{Make traffic light power plot}
\usage{
traffic_light_power_plot(power_plot, assumptions)
}
\arguments{
\item{power_plot}{Facet grid \code{power_plot} object}

\item{assumptions}{Assumptions regarding the likelihood of the included scenarios
in the \code{power_plot}. Make assumptions from upper-left strip to lower-right strip}
}
\value{
Returns a traffic light power plot
}
\description{
Traffic light coloring of the strips of a \code{power_plot}
}
\examples{

library(dplyr)

coteach_dat <- VWB22_pilot
dat_kjsigma2j <- dplyr::select(coteach_dat, kj, sigma2j = vg_ms_mean)


power_CHE_RVE_color_plot <-
  power_plot(
    J = seq(50, 100, 10),
    tau2 = c(0, 0.05, 0.1, 0.2)^2,
    omega2 = c(0.05, 0.15, 0.25, .35)^2,
    beta = 0.1,
    rho = c(.2, .4, .7, .9),
    model = "CHE",
    var_df = "RVE",
    pilot_data_kjsigma2 = dat_kjsigma2j,
    expected_studies = c(66, 86),
    color = TRUE,
    color_brewer = TRUE,
    iterations = 1,
    seed = 10052510
  )

traffic_light_power_plot(
  power_plot = power_CHE_RVE_color_plot,
  assumptions = c("unlikely", "likely", "expected", "likely",
                  "unlikely", "likely", "expected", "unlikely")
)

}
